# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Open4Printing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon/promo code at checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponInfo"
        "404":
          description: Invalid coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customer/register:
    post:
      operationId: customerRegister
      tags: [customers]
      summary: Create a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegisterBody"
      responses:
        "200":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerMe"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customer/login:
    post:
      operationId: customerLogin
      tags: [customers]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerMe"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /customer/me:
    get:
      operationId: customerMe
      tags: [customers]
      summary: Current customer session
      responses:
        "200":
          description: Current customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerMe"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customer/orders:
    get:
      operationId: listCustomerOrders
      tags: [customers]
      summary: Past orders for the signed-in customer
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customer/addresses:
    get:
      operationId: listCustomerAddresses
      tags: [customers]
      summary: Saved addresses for the signed-in customer
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedAddress"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCustomerAddress
      tags: [customers]
      summary: Save a new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressBody"
      responses:
        "201":
          description: Address saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedAddress"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List all products (admin)
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminProduct"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAdminProduct
      tags: [admin]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/products/{id}:
    patch:
      operationId: updateAdminProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAdminProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    Address:
      type: object
      properties:
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
      required: [line1, city, state, postalCode, country]
...
    ProductUploadConfig:
      type: object
      properties:
        artworkRequired: { type: boolean }
        allowsBackUpload: { type: boolean }
        preferredFormats:
          type: array
          items: { type: string }
        notes: { type: string }
      required: [artworkRequired, allowsBackUpload, preferredFormats, notes]
...
    OrderSummary:
      type: object
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        total: { type: number }
        paymentStatus: { type: string }
        orderStatus: { type: string }
        itemCount: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, orderNumber, email, firstName, lastName, total, paymentStatus, orderStatus, itemCount, createdAt]
...
    AdminProduct:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        categorySlug: { type: string }
        shortDescription: { type: string }
        description: { type: string }
        startingPrice: { type: number }
        enabled: { type: boolean }
        uploadConfig: { $ref: "#/components/schemas/ProductUploadConfig" }
        options:
          type: ["object", "null"]
          additionalProperties: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, slug, name, categorySlug, shortDescription, description, startingPrice, enabled, uploadConfig, createdAt, updatedAt]
    CreateProductBody:
      type: object
      properties:
        slug: { type: string }
        name: { type: string }
        categorySlug: { type: string }
        shortDescription: { type: string }
        description: { type: string }
        startingPrice: { type: number }
        enabled: { type: boolean }
        uploadConfig:
          oneOf:
            - { $ref: "#/components/schemas/ProductUploadConfig" }
            - { type: "null" }
        options:
          type: ["object", "null"]
          additionalProperties: true
      required: [slug, name, categorySlug, startingPrice]
    UpdateProductBody:
      type: object
      properties:
        name: { type: ["string", "null"] }
        categorySlug: { type: ["string", "null"] }
        shortDescription: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        startingPrice: { type: ["number", "null"] }
        enabled: { type: ["boolean", "null"] }
        uploadConfig:
          oneOf:
            - { $ref: "#/components/schemas/ProductUploadConfig" }
            - { type: "null" }
        options:
          type: ["object", "null"]
          additionalProperties: true
...
    ValidateCouponBody:
      type: object
      properties:
        code: { type: string }
        subtotal: { type: number }
      required: [code, subtotal]
    CouponInfo:
      type: object
      properties:
        code: { type: string }
        description: { type: string }
        percentOff: { type: ["integer", "null"] }
        amountOff: { type: ["number", "null"] }
        discountAmount: { type: number }
      required: [code, description, discountAmount]
    CustomerRegisterBody:
      type: object
      properties:
        email: { type: string }
        password: { type: string, minLength: 8 }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: ["string", "null"] }
      required: [email, password, firstName, lastName]
    CustomerLoginBody:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    CustomerMe:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: ["string", "null"] }
      required: [id, email, firstName, lastName]
    SavedAddress:
      type: object
      properties:
        id: { type: integer }
        label: { type: string }
        address: { $ref: "#/components/schemas/Address" }
        isDefault: { type: boolean }
      required: [id, label, address, isDefault]
    CreateAddressBody:
      type: object
      properties:
        label: { type: string }
        address: { $ref: "#/components/schemas/Address" }
        isDefault: { type: boolean }
      required: [label, address]
...