# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open4Printing API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProductBodyOptions } from "./createProductBodyOptions";
import type { ProductUploadConfig } from "./productUploadConfig";

export interface CreateProductBody {
  slug: string;
  name: string;
  categorySlug: string;
  shortDescription?: string;
  description?: string;
  startingPrice: number;
  enabled?: boolean;
  uploadConfig?: ProductUploadConfig | null;
  /** @nullable */
  options?: CreateProductBodyOptions;
}